+++ tests/test_snippet_fmt.py
@@ -1,12 +1,12 @@
 # stdlib
-import re
+import sys
 from typing import Dict, List, Union, no_type_check
 
 # 3rd party
 import dom_toml
 import pytest
 from _pytest.capture import CaptureResult
-from coincidence import AdvancedDataRegressionFixture, AdvancedFileRegressionFixture
+from coincidence import AdvancedDataRegressionFixture, AdvancedFileRegressionFixture, min_version
 from coincidence.params import param
 from consolekit.terminal_colours import strip_ansi
 from consolekit.testing import CliRunner, Result
@@ -209,3 +209,119 @@ def check_out(
 			}
 
 	advanced_data_regression.check(data_dict)
+
+
+def test_retval(
+		tmp_pathplus: PathPlus,
+		advanced_file_regression: AdvancedFileRegressionFixture,
+		advanced_data_regression: AdvancedDataRegressionFixture,
+		capsys,
+		):
+
+	(tmp_pathplus / "formate.toml").write_text((source_dir / "example_formate.toml").read_text())
+
+	config: SnippetFmtConfigDict = {"languages": {"python": {"reformat": True}}, "directives": ["code-block"]}
+
+	(tmp_pathplus / "demo.rst").write_lines([
+			".. code-block:: python",
+			'',
+			'\tprint    (\t"hello world"',
+			"\t)",
+			'',
+			'',
+			".. code-block:: python3",
+			'',
+			"\tdef for while if",
+			'',
+			])
+
+	with in_directory(tmp_pathplus):
+		assert reformat_file(tmp_pathplus / "demo.rst", config) == 1
+		advanced_file_regression.check_file(tmp_pathplus / "demo.rst")
+
+		result = capsys.readouterr()
+		stdout = result.out
+		stderr = result.err
+
+		advanced_data_regression.check({
+				"out": strip_ansi(stdout.replace(tmp_pathplus.as_posix(), "...")).split('\n'),
+				"err": strip_ansi(stderr.replace(tmp_pathplus.as_posix(), "...")).split('\n'),
+				})
+
+	with in_directory(tmp_pathplus):
+		assert reformat_file(tmp_pathplus / "demo.rst", config) == 0
+		advanced_file_regression.check_file(tmp_pathplus / "demo.rst")
+		assert capsys.readouterr().out == ''
+
+
+def test_retval_syntaxerror(
+		tmp_pathplus: PathPlus,
+		advanced_file_regression: AdvancedFileRegressionFixture,
+		advanced_data_regression: AdvancedDataRegressionFixture,
+		capsys,
+		):
+
+	(tmp_pathplus / "formate.toml").write_text((source_dir / "example_formate.toml").read_text())
+
+	config: SnippetFmtConfigDict = {"languages": {"python3": {"reformat": False}, }, "directives": ["code-block"]}
+
+	content = [
+			".. code-block:: python3",
+			'',
+			"\tdef for while if",
+			'',
+			".. code-block:: python3",
+			'',
+			'\tprint    (\t"hello world"',
+			"\t)",
+			'',
+			]
+
+	(tmp_pathplus / "demo.rst").write_lines(content)
+
+	data = {}
+
+	with in_directory(tmp_pathplus):
+		assert reformat_file(tmp_pathplus / "demo.rst", config) == 4
+		assert (tmp_pathplus / "demo.rst").read_lines() == content
+
+		result = capsys.readouterr()
+		stdout = result.out
+		stderr = result.err
+
+		data["error"] = {
+				"out": strip_ansi(stdout.replace(tmp_pathplus.as_posix(), "...")).split('\n'),
+				"err": strip_ansi(stderr.replace(tmp_pathplus.as_posix(), "...")).split('\n'),
+				}
+
+	config["languages"]["python3"]["reformat"] = True
+
+	with in_directory(tmp_pathplus):
+		assert reformat_file(tmp_pathplus / "demo.rst", config) == 5
+		advanced_file_regression.check_file(tmp_pathplus / "demo.rst")
+
+		result = capsys.readouterr()
+		stdout = result.out
+		stderr = result.err
+
+		data["reformat"] = {
+				"out": strip_ansi(stdout.replace(tmp_pathplus.as_posix(), "...")).split('\n'),
+				"err": strip_ansi(stderr.replace(tmp_pathplus.as_posix(), "...")).split('\n'),
+				}
+
+	if sys.version_info >= (3, 8):
+		# Offsets are wrong on earlier versions.
+		advanced_data_regression.check(data)
+
+	with in_directory(tmp_pathplus):
+		assert reformat_file(tmp_pathplus / "demo.rst", config) == 4
+		advanced_file_regression.check_file(tmp_pathplus / "demo.rst")
+
+		result = capsys.readouterr()
+		stdout = result.out
+		stderr = result.err
+
+		assert data["error"] == {
+				"out": strip_ansi(stdout.replace(tmp_pathplus.as_posix(), "...")).split('\n'),
+				"err": strip_ansi(stderr.replace(tmp_pathplus.as_posix(), "...")).split('\n'),
+				}